requirements.txt
pynput==1.7.6
google-generativeai==0.3.2
pyperclip==1.8.2
python-dotenv==1.0.0